requirements.txt
feedparser
requests
python-dotenv
mysql-connector-python==9.1.0